Excersice 1  code commit with some alteration

+++ Vijayvarsan/src/main/java/in/co/nmsworks/excercise/Addition.java
@@ -8,8 +8,8 @@ public class Addition {
     public static void main(String[] args) {
     Scanner scan = new Scanner(System.in);
         System.out.println("ENTER TWO VALUE FOR ADDITION ");
-        double value1 = scan.nextDouble();
-        double value2 = scan.nextDouble();
+        int value1 = scan.nextInt();
+        int value2 = scan.nextInt();
         System.out.println("ADDED VALUE IS = "+(value1+value2));
 }
 }
